datosUsuario renamed to addUsuario Style updated
